requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
boto3==1.26.46
botocore==1.29.46
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
lxml==4.9.2
outcome==1.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-environ==0.4.54
requests==2.28.1
s3transfer==0.6.0
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wsproto==1.2.0
